fix: renamed postHelpers to post
